+++ packages/jest-haste-map/src/__tests__/index.test.js
@@ -72,14 +72,13 @@ jest.mock('../crawlers/watchman', () =>
 const mockWatcherConstructor = jest.fn(root => {
   const EventEmitter = require('events').EventEmitter;
   mockEmitters[root] = new EventEmitter();
-  mockEmitters[root].close = jest.fn(callback => callback());
+  mockEmitters[root].close = jest.fn();
   setTimeout(() => mockEmitters[root].emit('ready'), 0);
   return mockEmitters[root];
 });
 
-jest.mock('sane', () => ({
-  NodeWatcher: mockWatcherConstructor,
-  WatchmanWatcher: mockWatcherConstructor,
+jest.mock('chokidar', () => ({
+  watch: jest.fn((patten, opts) => mockWatcherConstructor(opts.cwd)),
 }));
 
 jest.mock('../lib/WatchmanWatcher', () => mockWatcherConstructor);